[UPDATE]- Fix some bug and create SecurityService

+++ src/main/java/com/uit/realestate/dto/category/CategoryDto.java
@@ -14,5 +14,5 @@ public class CategoryDto {
 
     private String name;
 
-    private Integer totalItem;
+    private Long totalItem;
 }
